\name{traceplot}
\alias{traceplot}
\title{Trace Plot of MCMC Samples}
\description{
  Creates a trace plot of the posterior samples for the causal effect
  estimate from an MR-CUE analysis.
}
\usage{
traceplot(bhatpoint)
}
\arguments{
  \item{bhatpoint}{Numeric vector of posterior samples for the causal effect.}
}
\value{A ggplot2 object.}
