\name{MR.CUE-package}
\alias{MR.CUE-package}
\alias{MR.CUE}
\docType{package}
\title{
\packageTitle{MR.CUE}
}
\description{
\packageDescription{MR.CUE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MR.CUE}
\packageIndices{MR.CUE}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{MR.CUE}

Maintainer: \packageMaintainer{MR.CUE}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
