\name{MR.CUE-package}
\alias{MR.CUE-package}
\alias{MR.CUE}
\docType{package}
\title{
\packageTitle{MR.CUE}
}
\description{
\packageDescription{MR.CUE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MR.CUE}
\packageIndices{MR.CUE}
The main user-facing functions are \code{\link{MRCUE}}, \code{\link{MRCUEIndep}},
and \code{\link{MRCUESim}} for Mendelian randomization analysis, and
\code{\link{ReadSummaryStat}} and \code{\link{EstRho}} for data preparation.
}
\author{
\packageAuthor{MR.CUE}

Maintainer: \packageMaintainer{MR.CUE}
}
\references{
Cheng, Q. et al. (2022) Mendelian randomization accounting for complex correlated horizontal pleiotropy while elucidating shared genetic etiology. Nature Communications 13, 6490.
}
\keyword{ package }
